# api/openapi.yaml
openapi: 3.0.3

info:
  title: Delta Task API
  description: |
    Task management contract for the Delta interview project.

    This file is the single source of truth for the whole monorepo:

      - `backend/`  generates Spring MVC interfaces + models from it (openapi-generator).
      - `frontend/` generates TypeScript types from it (openapi-typescript).

    Neither side hand-writes its transport types. Change this file, run
    `pnpm generate` and `mvn compile`, and both sides fail to compile until
    they agree with the contract again.
  version: 1.0.0
  license:
    name: Proprietary

servers:
  - url: http://localhost:8080
    description: Local backend
  - url: http://localhost:5173
    description: Local Vite dev server (proxies /api to the backend)

tags:
  - name: tasks
    description: Create, inspect, modify, complete and delete tasks.

paths:
  /api/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: 檢視任務 — list tasks, optionally filtered by completion state.
      description: |
        Filtering happens server-side rather than in the client so that a large
        task set never has to cross the wire just to be discarded by the UI.
      parameters:
        - name: completed
          in: query
          required: false
          description: Omit for all tasks; `true`/`false` to filter by state.
          schema:
            type: boolean
      responses:
        '200':
          description: The matching tasks, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

    post:
      tags: [tasks]
      operationId: createTask
      summary: 建立任務 — create a task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: The created task.
          headers:
            Location:
              description: URL of the created task.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationFailed'

  /api/tasks/{id}:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    get:
      tags: [tasks]
      operationId: getTask
      summary: 檢視任務 — fetch a single task.
      responses:
        '200':
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/TaskNotFound'

    put:
      tags: [tasks]
      operationId: updateTask
      summary: 修改任務 — replace a task's editable fields.
      # [AI assisted 003] 「改類別時從目標類別的計數器重新發號」這條規則，是與 AI 討論
      # 「兩個計數器獨立之後，沿用舊號會不會撞號」推導出來的，結論一併寫進本說明。
      description: |
        Deliberately a full replace of the editable fields (title, description,
        category, dueDate) rather than a partial patch: completion state is
        owned by `PATCH /api/tasks/{id}/completion`, so an edit form can never
        silently reopen or close a task as a side effect of saving a title.

        Moving a task to the other `category` re-issues its `sequence` from that
        category's counter, so `feat-0003` becomes e.g. `bug-0007` rather than
        colliding with an existing `bug-0003`. `id` is untouched, so URLs and
        in-flight requests survive the move.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: The updated task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '404':
          $ref: '#/components/responses/TaskNotFound'

    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: 刪除任務 — delete a task.
      responses:
        '204':
          description: Deleted; no body.
        '404':
          $ref: '#/components/responses/TaskNotFound'

  /api/tasks/{id}/completion:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    patch:
      tags: [tasks]
      operationId: setTaskCompletion
      summary: 將任務標記為已完成或未完成 — set completion state.
      description: |
        Takes an explicit boolean rather than toggling. A toggle derives the new
        state from whatever the server currently holds, so two rapid clicks (or
        two devices) race to an unpredictable result; an explicit target state
        is idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompletionRequest'
      responses:
        '200':
          description: The task in its new completion state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '404':
          $ref: '#/components/responses/TaskNotFound'

components:
  parameters:
    TaskId:
      name: id
      in: path
      required: true
      description: Task identifier.
      schema:
        type: string
        format: uuid

  responses:
    TaskNotFound:
      description: No task exists with that id.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ValidationFailed:
      description: The request body failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    # [AI assisted 003] 使用 AI 協助設計工單編號的契約形狀：id 維持 uuid 給機器用，
    # 人看的編號改由 category + sequence 這一對組成（兩類別各自計數），並新增 dueDate。
    TaskCategory:
      type: string
      description: |
        工單類別 — the kind of work a task represents.

        Drives the display number's prefix, the card's colour and the board's
        grouping. Deliberately a closed enum rather than free text: the two
        values are a fixed part of the domain, and an enum makes an unknown
        value a 400 at the boundary instead of an unrenderable card.
      enum: [bug, feat]
      example: bug

    Task:
      type: object
      description: |
        A task as stored by the backend.

        Two identifiers, on purpose. `id` is the machine one — routes, requests,
        cache keys — and never changes. The number a human reads off the card is
        the pair (`category`, `sequence`), rendered client-side as `bug-0007` /
        `feat-0003`. Splitting them means the human-readable number can follow a
        task's category without ever breaking a URL.
      required: [id, category, sequence, title, completed, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
          example: 3f1a7c2e-9b04-4f5d-8a11-6c2d5e0f7b31
        category:
          $ref: '#/components/schemas/TaskCategory'
        sequence:
          type: integer
          format: int32
          minimum: 1
          description: |
            Per-category serial, assigned by the server and never accepted from
            a client. `bug` and `feat` count independently, so a `bug-0001` and
            a `feat-0001` both exist: it is the pair (`category`, `sequence`)
            that is unique, not `sequence` on its own.

            Sent as a plain integer; zero-padding to four digits is a
            presentation concern and belongs to the client.
          example: 7
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Write the OpenAPI contract
        description:
          type: string
          maxLength: 2000
          nullable: true
          example: Cover all five task operations.
        completed:
          type: boolean
          example: false
        dueDate:
          type: string
          format: date
          nullable: true
          description: |
            Optional due date, without a time or zone: a task is due on a day,
            not at an instant, and a timestamp would make the same task look
            overdue in one timezone and not in another.

            No `overdue` flag ships with the task — that is
            `dueDate < today && !completed`, which the client already re-derives
            every midnight anyway, whereas a server-computed flag would go stale
            in the browser the moment the date rolled over.
          example: '2026-09-30'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateTaskRequest:
      type: object
      description: |
        Payload for creating a task; new tasks always start incomplete.

        Carries no `sequence`: the serial is drawn from the chosen category's
        counter by the server, so a client cannot mint a duplicate number.
      required: [title, category]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        category:
          $ref: '#/components/schemas/TaskCategory'
        dueDate:
          type: string
          format: date
          nullable: true

    UpdateTaskRequest:
      type: object
      description: |
        Payload for replacing a task's editable fields.

        A full replace, so an omitted `description` or `dueDate` clears it.
        `category` is required for the same reason: a form that can move a task
        between categories must state where it lands, and leaving it out would
        make "no change" and "unset" indistinguishable.
      required: [title, category]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
          nullable: true
        category:
          $ref: '#/components/schemas/TaskCategory'
        dueDate:
          type: string
          format: date
          nullable: true

    TaskCompletionRequest:
      type: object
      description: Target completion state for a task.
      required: [completed]
      properties:
        completed:
          type: boolean

    Problem:
      type: object
      description: RFC 9457 problem details, as produced by Spring's ProblemDetail.
      required: [status, title]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
          example: Not Found
        status:
          type: integer
          format: int32
          example: 404
        detail:
          type: string
          example: No task exists with id 3f1a7c2e-9b04-4f5d-8a11-6c2d5e0f7b31.
        instance:
          type: string
          format: uri-reference